\name{cMap2data-package}
\alias{cMap2data-package}
\alias{cMap2data}
\docType{package}
\title{
cMap2 Data Package
}
\description{
Data package to provide default ranked profiles, clusters and Cytoscape \href{http://www.cytoscape.org/}{http://www.cytoscape.org/} file format for the DrugVsDisease package. The data is from drug screenings in the connectivity map (version 2) which is a project based at the Broad Institute of MIT and Harvard in Cambridge, Massachusetts (USA).
}
\details{
\tabular{ll}{
Package: \tab cMap2data\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2012-09-20\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
Data loaded automatically in the DrugVsDisease package.
}
\author{
C. Pacini

Maintainer: Julio Saez-Rodriguez{saezrodriguez@ebi.ac.uk}
}
\references{
Drug Profiles:
\href{http://www.broadinstitute.org/cmap/}{http://www.broadinstitute.org/cmap/}
Disease Profiles:
\href{http://www.ncbi.nlm.nih.gov/geo/}{http://www.ncbi.nlm.nih.gov/geo/}
}

\keyword{ package }
\seealso{

}
\examples{
#load the gene Ranked lists:
data(drugRL)
}
