\name{druglabels}
\alias{druglabels}
\docType{data}
\title{
Drug name maps to Drug Bank
}
\description{
Data frame containing the names of the compounds in the DvDdata drug profiles, with corresponding names to search the DrugBank [1] database through its web browser.
}
\usage{data(druglabels)}
\details{

}
\source{
\href{http://www.broadinstitute.org/cmap/}{http://www.broadinstitute.org/cmap/}
}
\references{
[1]   DrugBank 3.0: a comprehensive resource for 'omics' research on drugs. Knox C, Law V, Jewison T, Liu P, Ly S, Frolkis A, Pon A, Banco K, Mak C, Neveu V, Djoumbou Y, Eisner R, Guo AC, Wishart DS. Nucleic Acids Res. 2011 Jan;39(Database issue):D1035-41.

[2] The Connectivity Map: Using Gene-Expression Signatures to Connect Small Molecules, Genes, and Disease. Lamb J et al. Science 29 September 2006: 313 (5795), 1929-1935. [DOI:10.1126/science.1132939] 
}
\examples{
data(druglabels)

}
\keyword{datasets}
