\name{genelist}
\alias{genelist}
\docType{data}
\title{
Gene List Vector
}
\description{
List of HUGO genes included in the reference data sets.
}
\usage{data(genelist)}
\details{
A character vector listing those genes in the reference data sets. Obtained using BioMart annotations of the Affymetrix HGU-133A, HGU-133A2 and HGU-133Plus2 platforms.
}
\source{

}
\references{
Durinck et al. (2009). Mapping identifiers for the integration of genomic datasets with the R/Bioconductor package biomaRt. Nature Protocols 4, 1184-1191.
}
\examples{
data(genelist)

}
\keyword{datasets}
